% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival_Functions.R
\name{AdjacencyMatrix}
\alias{AdjacencyMatrix}
\title{Make an adjacency matrix from an sf object}
\usage{
AdjacencyMatrix(reaches)
}
\arguments{
\item{reaches}{a sf object with LINESTRING geometry representing a stream network}
}
\value{
an adjacency matrix used by INLA for spatial relationships
}
\description{
Make an adjacency matrix from an sf object
}
