% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_shapeprep.R
\name{AssembleReddData}
\alias{AssembleReddData}
\title{Combines survey tracks, counts, and a network into a data frame}
\usage{
AssembleReddData(
  shape,
  georedds = NA,
  georedds.type = NA,
  georedds.coords = NA,
  reddcrs = "wgs84",
  tolerance.redds = 250,
  tolerance.surveys = 10,
  surveys,
  survey.redds = NA,
  dist.units
)
}
\arguments{
\item{shape}{A sf object with LINESTRING geometries, or a sfnetwork}

\item{georedds}{An optional dataframe with coordinates for each redd observation}

\item{georedds.type}{An optional name for the column with classification of redds}

\item{georedds.coords}{A character vector with the column names for lon/lat in georedds}

\item{reddcrs}{A crs for the coordinates in georedds}

\item{tolerance.redds}{A maximum distance for snapping redds to the shape}

\item{tolerance.surveys}{A maximum distance for snapping surveys to the shape (should be small)}

\item{surveys}{A sf object with LINESTRINGs representing surveys}

\item{survey.redds}{optional name of a column with survey counts}

\item{dist.units}{a character string for the units}
}
\value{
a dataframe with dates, counts, and effort for each surveyed location
}
\description{
This function takes an sfnetwork or sf shape, a set of surveys, and optionally a set of
georeferenced redds observations, and returns a data set with the effort assigned to each reach
and the count types associated with that observation date.
}
\details{
Error Note: can cause reach numbers to be misassigned under some circumstances
Need to investigate further and then fix. In meantime, always be sure to check outputs.
}
