% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_shapeprep.R
\name{AssignCounts}
\alias{AssignCounts}
\title{Combines survey tracks and stream network into dataframe of counts.}
\usage{
AssignCounts(
  shape,
  surveys,
  effort,
  survey.counts,
  geodata = NA,
  geodata.counts,
  geodata.crs,
  geodata.coords,
  tolerance = 250,
  key,
  format.output = F
)
}
\arguments{
\item{shape}{A sf object with LINESTRING geometries, or a sfnetwork}

\item{surveys}{A sf object with LINESTRINGs representing surveys}

\item{effort}{dataframe of effort from the AssignEffort function}

\item{survey.counts}{optional name of a column with survey counts}

\item{geodata}{An optional dataframe with coordinates for each observation}

\item{geodata.counts}{character; name of the column with count for each observation; if missing assumes 1 count per line}

\item{geodata.crs}{A crs for the coordinates in geodata}

\item{geodata.coords}{A character vector with the column names for lon/lat in georedds}

\item{tolerance}{A maximum distance for snapping observations to the shape}

\item{key}{character string; optional name of a column present in both surveys and geodata, used for matching}

\item{format.output}{logical; the Format Counts function can usually be run using the same arguments, skipping an extra function call}
}
\value{
a dataframe with dates, counts, and effort for each surveyed location
}
\description{
This function takes an sfnetwork or sf shape, a set of surveys, an effort matrix and
optionally a set of georeferenced observations, and returns a dataframe with each survey
and the counts assigned to each reach associated with that survey
}
\details{
Error Note: can cause reach numbers to be mis-assigned under some circumstances
Need to investigate further and then fix. In meantime, always be sure to check outputs.
}
