% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_shapeprep.R
\name{AssignEffort}
\alias{AssignEffort}
\title{AssignEffort}
\usage{
AssignEffort(shape, surveys, dist.units, tolerance = 10, key)
}
\arguments{
\item{shape}{A sf object with LINESTRING geometries, or a sfnetwork}

\item{surveys}{A sf object with LINESTRINGs representing surveys}

\item{dist.units}{a character string for the units}

\item{tolerance}{A maximum distance for snapping surveys to the shape (should be small)}

\item{key}{character string; optional name of a column present in both surveys and georedds, used for matching}
}
\description{
Matches surveys to reaches and computes the level of effort associated with each
reach from each survey
}
