% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_shapeprep.R
\name{AssignReaches}
\alias{AssignReaches}
\title{Divide a network into reaches}
\usage{
AssignReaches(network, targetsize = 100, setnodes)
}
\arguments{
\item{network}{A sfnetwork}

\item{targetsize}{a desired size to}

\item{setnodes}{not working yet}
}
\value{
A new sfnetwork object divided into reaches of approximately the correct size
}
\description{
This function takes a network and divides it into distinct reaches based on a target size
if you want only set boundaries, use setnodes and set targetsize very high
The setnodes argument can be used to ensure certain reach boundaries (not implemented yet)
}
