% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_shapeprep.R
\name{AttachData}
\alias{AttachData}
\title{Attach Data to a network}
\usage{
AttachData(shape, refshape, fields = "all", logical = F, tolerance = 100)
}
\arguments{
\item{shape}{A sfnetwork or sf object with LINESTRING geometry to attach data to}

\item{refshape}{A sf object with LINESTRING geometry that contains the desired data}

\item{fields}{A character vector with the names of the columns to attach, or "all"}

\item{logical}{T/F determining how to handle logical values}

\item{tolerance}{A number used when matching lines that are not perfectly aligned}
}
\value{
A sfnetwork consisting of the original shape with data attached now
}
\description{
Take data contained in one sf LINESTRING object and attaches it to a network.
Values are averaged or combined where appropriate
}
