% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_shapeprep.R
\name{CheckNetwork}
\alias{CheckNetwork}
\title{Check a network and correct common problems}
\usage{
CheckNetwork(network, root)
}
\arguments{
\item{network}{A sfnetwork object}

\item{root}{A sf POINT or vector with coordinates for the root}
}
\value{
A sfnetwork with some additional data attached
}
\description{
this function checks that the direction of the stream network is away from the root
and fixes any problems, also adds a column named "parent" to the shape
}
