% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_shapeprep.R
\name{CropArea}
\alias{CropArea}
\title{Crop an sf shape - A graphical wrapper for the st_crop function}
\usage{
CropArea(shape, guides, xmin, xmax, ymin, ymax)
}
\arguments{
\item{shape}{A sf object}

\item{guides}{Any sf object you would like highlighted to help make your selections}

\item{xmin}{numeric; a minimum value for the left boundary}

\item{xmax}{numeric; a max value for the right boundary}

\item{ymin}{numeric; a minimum value for the bottom boundary}

\item{ymax}{numeric; a max value for the top boundary}
}
\value{
A sf object cropped to the user inputs
}
\description{
Crop an sf shape - A graphical wrapper for the st_crop function
}
