% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_VASTutility.R
\name{Dharmaplot}
\alias{Dharmaplot}
\title{Plot the dHARMa scaled residuals, accounts for dummy data}
\usage{
Dharmaplot(streamvast, span = 0.1)
}
\arguments{
\item{streamvast}{A streamvast object with fitted model and predictions}

\item{span}{A smoothing parameter for the line}
}
\value{
a plot mimicing the style of dHARMa, but accounts for dummy data
}
\description{
Plot the dHARMa scaled residuals, accounts for dummy data
}
