% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_shapeprep.R
\name{EditFeatures}
\alias{EditFeatures}
\title{Edit a shape or network in real time}
\usage{
EditFeatures(shape, root, guide)
}
\arguments{
\item{shape}{A sf or sfnetworks object to be editted}

\item{root}{A sf POINT for the root (might be obsolete)}

\item{guide}{An sf object that is plotted to help the user identify important shapes}
}
\value{
A sf or sfnetworks object with the user specificed alterations
}
\description{
Allows the user to add points or remove features by selecting them from the plot window
and command line prompts
}
