% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival_Functions.R
\name{EscapeReport}
\alias{EscapeReport}
\title{Make a dataframe with outputs from the VAST model for reporting}
\usage{
EscapeReport(streamvast, watershed)
}
\arguments{
\item{streamvast}{a streamvast object with completed predictions}

\item{watershed}{character; a name for the watershed}
}
\description{
Make a dataframe with outputs from the VAST model for reporting
}
