% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival_Functions.R
\name{EscapeReport2}
\alias{EscapeReport2}
\title{Make a table with relevant outputs from the VAST and survival models}
\usage{
EscapeReport2(streamvast, survival, escape, watershed)
}
\arguments{
\item{streamvast}{A streamvast}

\item{survival}{A table of survival values; usually from SurvivalTable}

\item{escape}{A table of escapement values, usually from MakeEscapement}

\item{watershed}{character; a name for the watershed}
}
\description{
Make a table with relevant outputs from the VAST and survival models
}
