% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_shapeprep.R
\name{ExtractBasin}
\alias{ExtractBasin}
\title{Extracts a basin or subbasin}
\usage{
ExtractBasin(shape, root, basin, exclude)
}
\arguments{
\item{shape}{an sf object consisting of linestrings (not multilinestrings!)}

\item{root}{a sf point, or length 2 vector with xy coordinates}

\item{basin}{a name or sf feature indicating the basin to be extracted}

\item{exclude}{vector indexes to be dropped, will attempt to match a single name}
}
\value{
A sfnetwork containing the specified subbasin
}
\description{
this function takes a shape and extracts the network above that node
Note that this function does not preserve information associated with the original shapefile
Future versions should allow the user to specify information to preserve, and allow for
}
