% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_shapeprep.R
\name{FindRoots}
\alias{FindRoots}
\title{FindRoots}
\usage{
FindRoots(shape, mainstem = NA)
}
\arguments{
\item{shape}{An sf object with linestring geometry and Identifier data}

\item{mainstem}{a name or unique id for the mainstem that basins branch from}
}
\value{
An sf object with points for each potential root
}
\description{
This function locates all vertices connected to the shape specified in mainstem
}
