% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_shapeprep.R
\name{FixLoops}
\alias{FixLoops}
\title{Guides the user in how to fix loops}
\usage{
FixLoops(network, root, guides = NULL)
}
\arguments{
\item{network}{A sfnetwork}

\item{root}{A sf object or coordinates for the root}

\item{guides}{optional list of sf objects that may assist decision making}
}
\value{
A sfnetwork with user specified sections removed
}
\description{
This function will guide the user through eliminating any loops in the network
using the command line. Note that for vast, a rooted tree is required,
so no reach can have multiple parents
}
