% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_shapeprep.R
\name{FormatCounts}
\alias{FormatCounts}
\title{Formats counts and effort along with spatial data for use in models}
\usage{
FormatCounts(countdata, effort, shape, merge = T, key)
}
\arguments{
\item{countdata}{dataframe of counts from the AssignCounts function}

\item{effort}{dataframe of effort from the AssignEffort function}

\item{shape}{A sf object with LINESTRING geometries, or a sfnetwork}

\item{merge}{logical; should observations of the same reach on the same date be combined (recommended)}

\item{key}{character; optional; name of a column present in both countdata and effort, used for matching}
}
\description{
This function combines the output of the EffortMatrix and AssignCounts functions
}
