% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_VASTutility.R
\name{FormatStreamData}
\alias{FormatStreamData}
\title{Final formatting of Stream data for VAST}
\usage{
FormatStreamData(counts, reaches, surveys, unit.conv = 0.0003048)
}
\arguments{
\item{counts}{A data frame with count data, presumably from AssembleReddData, or a filename}

\item{reaches}{A sf lines object, presumably from other StreamVAST functions, or a filename}

\item{surveys}{An optional set of surveys, presumably from MakeSurveyTracks, or a filename}

\item{unit.conv}{A conversion used for distance units, can set to 1}
}
\value{
A list with several elements necessary for VAST
}
\description{
Provides some final formatting and convenience steps before using VAST
}
