% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_VASTutility.R
\name{Jeremy_sample_variable}
\alias{Jeremy_sample_variable}
\title{Title}
\usage{
Jeremy_sample_variable(
  obj,
  newdata,
  what = "mu_g",
  n_samples = 100,
  sample_fixed = TRUE,
  seed = 123456,
  bias.correct = F
)
}
\arguments{
\item{obj}{a tinyVAST model object}

\item{newdata}{a dataset for making new predictions}

\item{what}{what predictions are desired, almost always "mu_g"}

\item{n_samples}{integer of samples}

\item{sample_fixed}{logical; almost always T}

\item{seed}{integer; a seed for the random draws}

\item{bias.correct}{logical; should the transformation bias correction be applied; very slow}
}
\value{
an object with simulated predictions on newdata
}
\description{
Title
}
