% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_VASTutility.R
\name{LoadStreamData}
\alias{LoadStreamData}
\title{Loads Stream data for VAST}
\usage{
LoadStreamData(watershed, size, unit.conv = 0.0003048)
}
\arguments{
\item{watershed}{A name for the folders and files}

\item{size}{a size used to identify the correct files}

\item{unit.conv}{A conversion used for distance units, can set to 1}
}
\value{
A list with several elements necessary for VAST
}
\description{
Loads and does some basic processing for previously saved data;
Unlikely to be generally useful
}
