% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_shapeprep.R
\name{LocateFeature}
\alias{LocateFeature}
\title{Locate a feature on a shape or network}
\usage{
LocateFeature(shape, type)
}
\arguments{
\item{shape}{A sf object with LINESTRING geometry or a sfnetwork}

\item{type}{Either "POINT", "VERTEX" or "SEGMENT" which indicates the type of shape to select}
}
\value{
A sf object containing the POINTS or LINESTRINGS selected by the user
}
\description{
Allows the user to select a number of shapes from a sf object or network by clicking
in the plot window. A shape will be converted to a sfnetwork first
}
