% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_shapeprep.R
\name{LocateRoot}
\alias{LocateRoot}
\title{Locate the root of a network}
\usage{
LocateRoot(network)
}
\arguments{
\item{network}{A sf network object with a valid root}
}
\value{
A sf object with POINT geometry representing the root of the network
}
\description{
Locates the root of a network, assuming it has one. If the network has unconnected nodes
or does not have a valid root, returns an error
}
