% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival_Functions.R
\name{MakeEscapement}
\alias{MakeEscapement}
\title{Make a table of escapement values based on the AUC method}
\usage{
MakeEscapement(
  streamvast,
  fixed.survival = NA,
  mult = 1,
  years = "all",
  reaches = "all"
)
}
\arguments{
\item{streamvast}{a fitted streamvast object with auc predictions}

\item{fixed.survival}{numeric; sets survival to a fixed value for all years/reaches}

\item{mult}{numeric: a value to multiply the number of unique redds by, typically a spawner/redds ratio}

\item{years}{integer; a vector of years to calculate escapement for}

\item{reaches}{integer; a vector fo reach numbers to calculate escapement for}
}
\value{
a table of escapement values by year and reach
}
\description{
Make a table of escapement values based on the AUC method
}
