% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_VASTutility.R
\name{MakePredictionData}
\alias{MakePredictionData}
\title{Make Prediction Data}
\usage{
MakePredictionData(streamvast, timespan = 15)
}
\arguments{
\item{streamvast}{a streamvast object with a fitted model}

\item{timespan}{a temporal span in days to space predictions for each season}
}
\value{
a dataframe suitable for predict function
}
\description{
Make Prediction Data
}
