% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival_Functions.R
\name{MakeReddSurvival}
\alias{MakeReddSurvival}
\title{Convert redd data into survival data set}
\usage{
MakeReddSurvival(
  streamvast,
  redds,
  redd.ids = "redd_name_txt",
  redd.crs = "wgs84",
  survey.tol = 250,
  redd.status = "redd_status_code",
  min.surveys = 1,
  buffer = 14
)
}
\arguments{
\item{streamvast}{A streamvast object with reachdata, countdata, and surveydata}

\item{redds}{a dataframe containing observations of redds in each row}

\item{redd.ids}{character; the name of the column with consistent id for each read over time}

\item{redd.crs}{a crs for the redds}

\item{survey.tol}{numeric; a max tolerance for matching redds to survey lines}

\item{redd.status}{character; the name of the column with redd status (should use NR, SV, NV coding)}

\item{min.surveys}{A minimum number of surveys for observations of a reach to be included}

\item{buffer}{A number of days to be added to redds found on the first or last day of a season}
}
\value{
a dataframe formated for use with INLA survival models
}
\description{
This function takes a dataframe of repeated redd observations over time and converts them into
a format suitable for modeling redd survival.
}
