% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_shapeprep.R
\name{Makesfnetwork}
\alias{Makesfnetwork}
\title{Make a sf network}
\usage{
Makesfnetwork(shape, attach.data = F, simple = T)
}
\arguments{
\item{shape}{An sf object with LINESTRING or MULTILINESTRING geometry}

\item{attach.data}{Logical, whether data fields should be preserved}

\item{simple}{Logical, whether to use the fast and simple method, or the slower more thorough method}
}
\value{
A sfnetwork object
}
\description{
This function converts a sf shape with linestring or multilinestring geometries into
a sfnetwork object. Mostly, it is used internally.
}
