% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_shapeprep.R
\name{MakeSurveyTracks}
\alias{MakeSurveyTracks}
\title{Makes sf LINeSTRINGs representing survey transects}
\usage{
MakeSurveyTracks(
  shape,
  surveys,
  surveys.crs = "wgs84",
  save.col = "all",
  maxdist = 25,
  survey.coords = c("startlon", "startlat", "endlon", "endlat")
)
}
\arguments{
\item{shape}{A sf network or sf object with LINESTRING geometry}

\item{surveys}{A data frame with coordinates and any other desired information}

\item{surveys.crs}{a crs object or code for the survey coordinates}

\item{save.col}{character vector with names of any columns to save, or "all"}

\item{maxdist}{a maximum distance for when the coordinates are compared to the shape}

\item{survey.coords}{vector of names or column numbers; order matters}
}
\value{
A sf object with LINESTRINGS and data representing each survey
}
\description{
This function uses coordinates to locate the start and stop points for a survey
and then draws a line that follows the stream contours between those points.
If multiple paths are available, it will choose the shortest, so it's best to eliminate
multichannel sections ahead of time
}
