% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_VASTutility.R
\name{MakeVASTData}
\alias{MakeVASTData}
\title{This function takes data from the Load Stream data function and formats it for use with VAST}
\usage{
MakeVASTData(
  data,
  reachdata,
  unit.conv = 1,
  reddname = "Redds",
  startdate = NA,
  enddate = NA,
  padzero = T,
  Time = "Year"
)
}
\arguments{
\item{data}{a data frame with time, space, effort, and observations}

\item{reachdata}{A sf object, presumably from LoadStreamData}

\item{unit.conv}{a unit conversion for distances}

\item{reddname}{the name in data that corresponds to the Redd count}

\item{startdate}{dates to truncate or expand the data in yyyy-mm-dd format}

\item{enddate}{dates to truncate or expand the data in yyyy-mm-dd format}

\item{padzero}{should survey with zero observations be added at start/end dates}

\item{Time}{character giving the timescale to use}
}
\description{
This function takes data from the Load Stream data function and formats it for use with VAST
}
