% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival_Functions.R
\name{plotEscapement}
\alias{plotEscapement}
\title{Make a plot of yearly escapement values}
\usage{
plotEscapement(
  streamvast,
  obs.escape,
  title,
  ribbons = NA,
  median = T,
  years = "all",
  reaches = "all",
  color = 2
)
}
\arguments{
\item{streamvast}{A streamvast object with escapement predictions}

\item{obs.escape}{A vector of observed or reported escapement values; must be equal to the # of years in escape}

\item{title}{character; a title to be passed to ggtitle}

\item{ribbons}{a positive integer, the number of shaded ribbons for the CI}

\item{median}{should the plot show the predicted escapement based on the mle or the median of the posterior}

\item{years}{a vector of years to be included}

\item{reaches}{a vector of reach numbers to be included}

\item{color}{integer or character; a color for the plot}
}
\value{
either a ggplot object or a dataframe depending on the 'plot' option
}
\description{
Make a plot of yearly escapement values
}
