% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_VASTutility.R
\name{plotPredictionMap}
\alias{plotPredictionMap}
\title{Maps, the desired data, with options for facetting}
\usage{
plotPredictionMap(
  data,
  mapvar = "preds",
  reaches,
  facet = NA,
  reach.names = "reachid",
  make.labels = T,
  xaxis.breaks = NA,
  yaxis.breaks = NA,
  palette = "turbo"
)
}
\arguments{
\item{data}{a data frame with predictions and any grouping data}

\item{mapvar}{the column in data to map}

\item{reaches}{an sf object with reach data}

\item{facet}{a column to determine the facets, or a vector of values}

\item{reach.names}{the column in reaches that holds the id numbers}

\item{make.labels}{should the reaches be labelled, turn off for multiple years}

\item{xaxis.breaks}{Vector of x values for axis lines; useful to reduce crowding}

\item{yaxis.breaks}{Vector of y values for axis lines; useful to reduce crowding}

\item{palette}{a viridis color palette for plotting}
}
\value{
A colored heatmap of the desired variable mapped over the stream network
}
\description{
Maps, the desired data, with options for facetting
}
