% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_VASTutility.R
\name{plotStream}
\alias{plotStream}
\title{Makes a line plot that shows tributaries by stream distance}
\usage{
plotStream(
  data,
  plotvar,
  streamdist = "streamdist",
  rnames = "STRM_NA",
  reaches,
  main = NA
)
}
\arguments{
\item{data}{a data frame with a variable to plot matched to reach}

\item{plotvar}{the column name in data to be plotted}

\item{streamdist}{name of a column in reaches with stream distance}

\item{rnames}{name or id of reaches}

\item{reaches}{a sf object containing the reach data}

\item{main}{the name of the mainstem or most important stream, which will be highlighted}
}
\value{
A plot of the variable with respect ot stream distance
}
\description{
Makes a line plot that shows tributaries by stream distance
}
