% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival_Functions.R
\name{plotSurvivalCurves}
\alias{plotSurvivalCurves}
\title{Plot survival curves from a INLA survival model}
\usage{
plotSurvivalCurves(model, data, year = "all", reach = "all", title)
}
\arguments{
\item{model}{a fitted INLA survival model}

\item{data}{a dataframe of redd observations, usually from MakeReddSurvival}

\item{year}{numeric; a subset of years to include in the plot}

\item{reach}{numeric; a subset of reach numbers to include in the plot}

\item{title}{character; a title to pass to ggtitle()}
}
\value{
a ggplot showing the
}
\description{
Plot survival curves from a INLA survival model
}
