% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_shapeprep.R
\name{PointSetup}
\alias{PointSetup}
\title{Point Setup and checking}
\usage{
PointSetup(point, crs)
}
\arguments{
\item{point}{A sf, sfc, dataframe, or vector containing geometry or coordinates representing a point}

\item{crs}{A crs to apply to the point, if it is not an sf or sfc object}
}
\value{
A sf object with POINT geometry
}
\description{
Checks that a point is an sf object with the appropriate crs, or converts inputs into an sf
object with POINT geometry; mostly used internally to check user inputs
}
