% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_shapeprep.R
\name{PruneNetwork}
\alias{PruneNetwork}
\title{Prunes a network based on various criteria}
\usage{
PruneNetwork(network, root, exclude, match, tolerance = 100, plot = T)
}
\arguments{
\item{network}{A sfnetwork to be pruned}

\item{root}{a sf point or coordinates for the root}

\item{exclude}{a logical vector or vector of values to exclude}

\item{match}{a sf object with points or lines}

\item{tolerance}{a tolerance for matching points to the network}

\item{plot}{logical, should a plot be made}
}
\value{
A sfnetwork pruned as specified
}
\description{
New and improved version of prune network, it handles a larger variety of inputs and is better implemented
exclude and match can be used in a variety of ways, and the function can sometimes handle them at the same time
But it is often better to call the function multiple times, using different arguments
}
