% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_shapeprep.R
\name{RiverMeasureLL}
\alias{RiverMeasureLL}
\title{Converts river measures and Lon/Lat coordinates}
\usage{
RiverMeasureLL(shape, measures, points, crs)
}
\arguments{
\item{shape}{A sf LINESTRING corresponding stream}

\item{measures}{a vector of river measures, in the same units as shape}

\item{points}{a dataframe of coordinates or sf object with POINT geometries}

\item{crs}{A crs for the points}
}
\value{
a sf object with POINT geometries and matched coordinates and river measure data
}
\description{
This tool take a LINESTRING and either returns coordinates based on a set of river measures
or river measures based on a set of given coordinates. Note that a single, unbranched LINESTRING
should be used for shape, which may require calling the function multiple times for complex systems
The shape should be constructed in the right direction as well
(this should happen automatically if using the other functions).
}
