% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_shapeprep.R
\name{RootNetwork}
\alias{RootNetwork}
\title{Adds or changes the root of a network}
\usage{
RootNetwork(network, root, root.crs, tolerance = Inf)
}
\arguments{
\item{network}{A sfnetwork}

\item{root}{A sf point, or coordinates for the new root}

\item{root.crs}{a crs or string for the root, if it is different from the network}

\item{tolerance}{A distance that determines whether to use existing nodes or blend in a new one}
}
\value{
A sfnetwork with the new root
}
\description{
This function sets up the root and ensures that all nodes are connected to it, and
preserves any old data. Any nodes not connected to the root are removed. By default, the
function will root the network at the closest existing node to the specified root. Set tolerance = 0
to instead cause the function to create a new node
}
