% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_VASTutility.R
\name{SetVastCovariates}
\alias{SetVastCovariates}
\title{Add habitat covariates to a streamvast object}
\usage{
SetVastCovariates(
  streamvast,
  pform,
  dform,
  pconfig = NULL,
  dconfig = NULL,
  spcovars,
  tempcovars,
  covariatedata,
  append = T,
  center = T,
  scale = T
)
}
\arguments{
\item{streamvast}{A streamvast object with temporal and spatial frames defined}

\item{pform}{A one-sided formula for covariates affecting the probability of occurance}

\item{dform}{A one-sided formula for covariates affecting density given presence}

\item{pconfig}{A data frame of control values for pform: See ?VAST::make_data for details}

\item{dconfig}{A data frame of control values for dform: See ?VAST::make_data for details}

\item{spcovars}{A dataframe or sf object with spatially referenced covariate data}

\item{tempcovars}{A dataframe with temporally referenced covariate data}

\item{covariatedata}{A dataframe with covariates matched to vastdata}

\item{append}{Logical; Should the covariates be appended to any pre-existing covariate data in the object}

\item{center}{Logical; Should the covariates be centered to a mean value of zero}

\item{scale}{Logical; Should the covariates be scaled to have standard deviation of one}
}
\value{
A streamvast object with covariate features appropriately formatted
}
\description{
Add habitat covariates to a streamvast object
}
