% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival_Functions.R
\name{SurvivalTable2}
\alias{SurvivalTable2}
\title{An Updated Survival Table}
\usage{
SurvivalTable2(model, newdata, extrayears, extramethod)
}
\arguments{
\item{model}{A fitted INLA survival model}

\item{extrayears}{any years not present in the data for which an estimate is desired}

\item{extramethod}{a function that will be applied to the data to extrapolate to new years}

\item{pred.index}{A vector of numbers that correspond to the table in the model data}
}
\description{
This version can use the pred index to generate predictions under specific
covariate values. This will probably be combined with SurivalTable after some
additional testing
}
