% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_VASTutility.R
\name{VASTeval}
\alias{VASTeval}
\title{Get predictions matched to observations from a vast model}
\usage{
VASTeval(model, data, effort = "Effort", time.table = NA)
}
\arguments{
\item{model}{A VAST model object}

\item{data}{The model data used to fit the VAST model}

\item{effort}{A name for the column with effort information}

\item{time.table}{An optional table matching time to a date, used to provide year, month, and day}
}
\value{
A data frame of observed densities and counts matched to predictions
}
\description{
Get predictions matched to observations from a vast model
}
