% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_VASTutility.R
\name{VASTpreds}
\alias{VASTpreds}
\title{Get density predictions from a VAST model}
\usage{
VASTpreds(
  streamvast,
  newdata,
  makeauc = T,
  nsims = 100,
  seed = 123456,
  bias.correct = F
)
}
\arguments{
\item{streamvast}{A VAST model object}

\item{newdata}{a data frame with all necessary columns over which to make predictions}

\item{makeauc}{logical; should auc measures be calculated. Can turn off to save time.}

\item{nsims}{interger; a number of simulated draws to make from the joint posterior}

\item{seed}{integer; a seed for random number generation}

\item{bias.correct}{logical; should the poisson transformation bias be corrected; very slow}
}
\value{
A streamvast object with a dataframe of predictions added
}
\description{
This function quickly constructs a data frame with VAST density predictions matched to
time and space. It also samples and provides upper and lower bounds (90\%)
}
