% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_VASTutility.R
\name{VASTtotals}
\alias{VASTtotals}
\title{Assemble estimates of total population index by time}
\usage{
VASTtotals(model, eval)
}
\arguments{
\item{model}{A fitted VAST model}

\item{eval}{A dataframe produced by VASTeval}
}
\value{
A data frame with the population index and CIs
}
\description{
Assemble estimates of total population index by time
}
