% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_VASTutility.R
\name{ConstructStreamVAST}
\alias{ConstructStreamVAST}
\title{ConstructStreamVAST
This function constructs objects with class "StreamVAST". Depending on the arguments supplied, it will either
initialize the object from countdata and reachdata, or it will add details to an already defined StreamVAST object.}
\usage{
ConstructStreamVAST(
  countdata,
  reachdata,
  surveydata,
  countname,
  reachname,
  unitconv
)
}
\arguments{
\item{countdata}{A dataframe containing count data, such as outputed from AssembleReddData}

\item{reachdata}{A sf object with data for reaches, such as from AssignReaches}

\item{surveydata}{A optional data frame or sf object with survey information, such as from MakeSurveyTracks, doesn't do much right now}

\item{countname}{A column name from countdata to use}

\item{reachname}{a column name from reachdata to use}

\item{StreamVAST}{An StreamVAST object to append details}
}
\value{
A streamvast object with appropriate formatting
}
\description{
ConstructStreamVAST
This function constructs objects with class "StreamVAST". Depending on the arguments supplied, it will either
initialize the object from countdata and reachdata, or it will add details to an already defined StreamVAST object.
}
