% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_shapeprep.R
\name{EditFeatures}
\alias{EditFeatures}
\title{Edit a shape or network in real time}
\usage{
EditFeatures(shape, root)
}
\arguments{
\item{shape}{A sf or sfnetworks object to be editted}

\item{root}{A sf POINT for the root (might be obsolete)}
}
\description{
Allows the user to add points or remove elements by selecting them from the plot window
and command line prompts
}
