% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_shapeprep.R
\name{ExtractBasin}
\alias{ExtractBasin}
\title{Extracts a basin or subbasin}
\usage{
ExtractBasin(
  shape,
  root,
  basins,
  exclude,
  extract.inner = F,
  attach.data = T,
  upper,
  upper.crs,
  root.tolerance = Inf
)
}
\arguments{
\item{shape}{an sf object consisting of linestrings (not multilinestrings!)}

\item{root}{a sf point, or length 2 vector with xy coordinates}

\item{basins}{a vector of names or sf feature indicating the basin to be extracted}

\item{exclude}{vector indexes to be dropped, will attempt to match a single name}

\item{extract.inner}{logical; should nodes that are in between the basins and the root be extracted}

\item{attach.data}{logical; should old data fields be obtained; time consuming}

\item{upper}{a point or points to stop the extraction}

\item{upper.crs}{a crs string for the upper point, if different from shape}

\item{root.tolerance}{a tolerance setting for the root. Low values will add the root to the network as a new node.}
}
\value{
A sfnetwork containing the specified subbasins
}
\description{
this function takes a shape and extracts the network above that node
Note that this function does not preserve information associated with the original shapefile
Future versions should allow the user to specify information to preserve, and allow for
}
