% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival_Functions.R
\name{MakeEscapement}
\alias{MakeEscapement}
\title{Make a table of escapement values based on the AUC method}
\usage{
MakeEscapement(streamvast, survival, fixed.survival = F, mult = 1.62)
}
\arguments{
\item{streamvast}{a fitted streamvast object with auc predictions}

\item{survival}{a table of survival values, usually from SurvivalTable}

\item{fixed.survival}{numeric; sets survival to a fixed value for all years/reaches}

\item{mult}{numeric: a value to multiply the number of unique redds by, typically a spawner/redds ratio}
}
\value{
a table of escapement values by year and reach
}
\description{
Make a table of escapement values based on the AUC method
}
