% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival_Functions.R
\name{plotEscapement}
\alias{plotEscapement}
\title{Make a plot of yearly escapement values}
\usage{
plotEscapement(escape, obs.escape, title, plot = T)
}
\arguments{
\item{escape}{A table of escapement values, usually from MakeEscapement}

\item{obs.escape}{A vector of observed or reported escapement values; must be equal to the # of years in escape}

\item{title}{character; a title to be passed to ggtitle}

\item{plot}{logical; should the function return the plot, or just the data}
}
\value{
either a ggplot object or a dataframe depending on the 'plot' option
}
\description{
Make a plot of yearly escapement values
}
