% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_VASTutility.R
\name{plotStream}
\alias{plotStream}
\title{Makes a line plot that shows tributaries by stream distance}
\usage{
plotStream(
  streamvast,
  plotvar,
  streamname,
  usepreds,
  title,
  show.names = "all"
)
}
\arguments{
\item{streamvast}{a streamvast object with predictions or evaluations}

\item{plotvar}{the column name in data to be plotted}

\item{streamname}{the column name in the reaches with the stream name}

\item{usepreds}{logical, should the prediction or evaluation data be used, if absent, the function will search}

\item{title}{a title to display above the graph}

\item{show.names}{character; a vector of names to highlight, useful for large watersheds}
}
\value{
A plot of the variable with respect ot stream distance
}
\description{
Makes a line plot that shows tributaries by stream distance
}
