% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_VASTutility.R
\name{SetTemporalFrame}
\alias{SetTemporalFrame}
\title{This function takes a streamvast object defines a temporal sampling frame and formats everything for use with vast}
\usage{
SetTemporalFrame(
  streamvast,
  startdate = NA,
  enddate = NA,
  Time = "Year",
  season,
  padyear = F,
  padhabitat = F
)
}
\arguments{
\item{streamvast}{a streamvast object, minimally with countdata, reachdata defined}

\item{startdate}{dates to truncate or expand the data in yyyy-mm-dd format}

\item{enddate}{dates to truncate or expand the data in yyyy-mm-dd format}

\item{Time}{character giving the timescale to use}

\item{season}{a integer vector of length with order start Month, start day, end month, end day for a 'year' or 'season'}

\item{padyear}{adds zeroes to the start and end of each runyear}

\item{padhabiat}{adds zeros at regular intervals to areas designated non-habitat}
}
\value{
A streamvast object with formated with the appropriate temporal frame
}
\description{
This function takes a streamvast object defines a temporal sampling frame and formats everything for use with vast
}
