% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_shapeprep.R
\name{SimplifyNetwork}
\alias{SimplifyNetwork}
\title{Simplifies a sfnetwork}
\usage{
SimplifyNetwork(network, tolerance = 1000, tolerance2 = 1, makeplot = T)
}
\arguments{
\item{network}{A sfnetwork}

\item{tolerance}{A integer, segments with no connections below this size are removed}

\item{tolerance2}{A fine scale tolerance for resolving overlapping edges; ought to be small}

\item{makeplot}{should the funciton output a plot showing before/after versions of the network}
}
\value{
A sfnetwork object
}
\description{
This function removes extraneous nodes and small segments from a network that are often
leftover after other operations
}
