% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival_Functions.R
\name{SurvivalTable}
\alias{SurvivalTable}
\title{Summarize the results of a survival model in tabular form}
\usage{
SurvivalTable(model, extrayears)
}
\arguments{
\item{model}{A fitted INLA survival model}

\item{extrayears}{any years not present in the data for which an estimate is desired}
}
\value{
an adjacency matrix used by INLA for spatial relationships
}
\description{
Summarize the results of a survival model in tabular form
}
