% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_VASTutility.R
\name{VASTeval}
\alias{VASTeval}
\title{Get predictions matched to observations from a vast model. Fills in the evaluation data set, and adds some columns
to the other slots.}
\usage{
VASTeval(streamvast)
}
\arguments{
\item{streamvast}{A VAST model object}
}
\value{
A streamvast object with an evaluation data set included
}
\description{
Get predictions matched to observations from a vast model. Fills in the evaluation data set, and adds some columns
to the other slots.
}
