% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_VASTutility.R
\name{VASTpreds}
\alias{VASTpreds}
\title{Get density predictions from a VAST model}
\usage{
VASTpreds(streamvast)
}
\arguments{
\item{streamvast}{A VAST model object}
}
\value{
A streamvast object with a dataframe of predictions added
}
\description{
This function quickly constructs a data frame with VAST density predictions matched to
time and space. It also samples and provides upper and lower bounds (90\%)
}
